+++ src/main/java/oit/is/work/team2/controller/AdminRoomsController.java
@@ -1,0 +1,5 @@
+package oit.is.work.team2.controller;
+
+public class AdminRoomsController {
+    
+}
